requirements.txt
annotated-types==0.5.0
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.11
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
confection==0.1.3
contourpy==1.1.1
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
exceptiongroup==1.1.3
fonttools==4.42.1
goose3==3.1.17
gTTS==2.3.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.0
outcome==1.2.0
packaging==23.1
pandas==2.1.3
pathy==0.10.2
Pillow==10.0.1
preshed==3.0.9
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.6.0/pt_core_news_sm-3.6.0-py3-none-any.whl#sha256=61a9193dd1cb7ae1a65a85f7c98f4b14077a8bd3dc34a486aa620a54b353a1d9
pyahocorasick==2.0.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.13.0
selenium==4.12.0
six==1.16.0
sklearn==0.0.post9
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.0
srsly==2.4.8
thinc==8.1.12
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wasabi==1.1.2
wordcloud==1.9.2
wsproto==1.2.0
